completed

+++ src/main/webapp/board1/board1_DetailView.jsp
@@ -49,7 +49,7 @@ div {
 	background-color: bisque;
 }
 
-.good_path:hover{
+.good_path:hover,.goodcancel_path:hover{
 	cursor:pointer;
 }
 </style>
@@ -88,13 +88,20 @@ div {
 			<div id="title" class="col-10">${dto.title }</div>
 			<div id="view_Count" class="col-2">${dto.view_count }</div>
 			<div id="img_Box" class="col-5">
-				<img src="files/${dto.fileName }" style="width: 100%; height: 100%;">
+				<c:choose>
+					<c:when test="${dto.fileName==null }">
+						<img src="/img/pig2.png" style="width: 100%; height: 100%;">
+					</c:when>
+					<c:otherwise>
+						<img src="files/${dto.fileName }" style="width: 100%; height: 100%;">
+					</c:otherwise>
+				</c:choose>
 			</div>
 			<div id="contents_Box" class="col-7">
 				<div id="item">${dto.item}</div>
 				<div id="contents">${dto.contents}</div>
 			</div>
-			<div id="item_price">전체 금액 : ${dto.item_price}</div>
+			<div id="item_price"></div>
 			<div>
 				<div id="graph" style="width: 70%; margin: auto;">
 					<div style="width:${agreeRatio}%;height:10px" id="agreeRatio"></div>
@@ -119,7 +126,7 @@ div {
 						</div>
 						<div style="border: 0px">
 							<input type="text" placehold="왜 그렇게 생각하나요?" style="width: 100%;"
-								name="contents" id="contents_reply">
+								name="contents" id="contents_reply" maxlength=300>
 						</div>
 					</div>
 					<div id="left" class="col-3" style="border: 0px">
@@ -130,6 +137,8 @@ div {
 				<hr>
 				</form>
 				<!-- --댓글 View------------------------------------------------------------------------------ -->
+				
+				
 				<c:forEach var="i" items="${list }">
 					<form action="/modify.brd1_reply">
 						<div class="replyViewcontainer" class="row">
@@ -144,10 +153,10 @@ div {
 							</div>
 							<div class="col-8" style="border:0px">
 								<div class="writer">작성자 : ${i.writer }</div>
-								<input type="text" name="reply_contents" value='${i.contents }' readonly> 
+								<input type="text" name="reply_contents" value='${i.contents }' readonly maxlength=300> 
 								<input type="hidden" name="replySeqToUpdate" value=${i.seq }>
 								<input type="hidden" name="parent_seq2" value=${dto.seq }>
-								<div>${i.write_date }</div>
+								<div id="w${i.seq }"></div>
 							</div>
 							<div class="col-2" style="border:0px">
 								<c:if test="${i.writer==nickname}">
@@ -191,9 +200,46 @@ div {
 
 	<script>
 	<!--  게시글 script-------------------------------------------------------------------- -->
-		
-
 		window.onload = function(){	
+			if(${nickname==null}){
+				alert("로그인 후 이용 가능합니다.")
+				location.href="/index.jsp"
+			}
+//          0. 날짜 변환(오늘날짜는 시간, 전날은 연,월,일 표시)
+			let today=new Date();//오늘날짜
+			
+			toYear = today.getFullYear();
+			toMonth = today.getMonth()+1;
+			toDate = today.getDate()
+
+			for(let i=0;i<${glist}.length;i++){
+				let seq = ${glist}[i].seq;
+				let write_date = new Date(${glist}[i].write_date);
+				
+				year = write_date.getFullYear();
+				month = write_date.getMonth()+1;
+				date = write_date.getDate();
+				hour = write_date.getHours();
+				minutes = write_date.getMinutes();
+				
+				if(toYear==year&&toMonth==month&&toDate==date){
+					write_date = hour +":"+minutes;
+					$("#w"+seq).text(write_date);
+				}else{
+					write_date = year +"."+month +"."+date;
+					$("#w"+seq).text(write_date);					
+				}
+			}
+			
+			let money=${dto.item_price}
+			let moneyForm = "총 금액 : "+money.toLocaleString() +"원"  
+			$("#item_price").text(moneyForm)
+			
+			
+			// 로드시 좋아요 한 댓글 검정색 처리	
+			if(${nickname==null}){
+				location.href="/index.jsp"
+			}
 			$.ajax({
 				url:"/isGoodChecked.brd1",
 				data:{
@@ -201,7 +247,6 @@ div {
 				},
 				dataType:"json"
 			}).done(function(res){
-				console.log(res);
 				$(res).each(function(){
 					$("#c"+this.reply_Seq).parent().css("display","block");
 					$("#"+this.reply_Seq).parent().css("display","none");
@@ -214,25 +259,41 @@ div {
 				$("#contents_reply").attr("readonly",true);
 				$("#contents_reply").attr("placeholder","댓글은 한 게시물 당 한 개만 작성이 가능합니다.");
 			}
+
 		}
-		
+		 
 		
 		$("#toList").on("click",function(){
 			location.href="/list.brd1?cpage=1";
 		})
 		$("#modify").on("click",function(){//수정하기
+			if(${nickname==null}){
+				alert("로그인 후 이용가능합니다.");
+				location.href="/index.jsp"
+			}
 			location.href="/toModifyForm.brd1?seq=${dto.seq}";
 		})
 		$("#delete").on("click",function(){
+			if(${nickname==null}){
+				alert("로그인 후 이용가능합니다.");
+				location.href="/index.jsp"
+			}
 			let result = confirm("정말 삭제하시겠습니까?")
 			if(result){
 				location.href="/delete.brd1?seq=${dto.seq}";
 			}
+
 		})
+		
+		
 	<!-- 댓글 script-------------------------------------------------------------------- -->
 
 	
-		$("#reply").on("click",function(){		
+		$("#reply").on("click",function(){	
+			if(${nickname==null}){
+				alert("로그인 후 이용가능합니다.");
+				location.href="/index.jsp"
+			}
 // 			1. 제출 시 radio 체크여부 확인
 			if(!$('input:radio[name="agree"]').is(":checked")){
 				alert("'승인','불가' 두가지 사항 중 선택해 주세요.");
@@ -247,6 +308,10 @@ div {
 	
 //			3. 댓글 수정 기능
 		$(".modify_btn").on("click",function(){
+			if(${nickname==null}){
+				alert("로그인 후 이용가능합니다.");
+				location.href="/index.jsp"
+			}
 			$(this).css('display','none');//수정 버튼 none처리
 			$(this).parent().parent().prev().children().eq(1).attr('readonly',false);
 			//댓글창 text창 활성화
@@ -289,13 +354,25 @@ div {
 		})
 //			4. 댓글 삭제 기능
 		$(".delete_btn").on("click",function(){
-			let seq = $(this).parent().parent().prev().children().eq(2).val();
-			let board_Seq = $(this).parent().parent().prev().children().eq(3).val(); 
-			location.href="/delete.brd1_reply?seq="+seq+"&board_Seq="+board_Seq;
+			if(${nickname==null}){
+				alert("로그인 후 이용가능합니다.");
+				location.href="/index.jsp"
+			}
+			
+			result = confirm("댓글을 정말로 삭제하시겠습니까?")
+			if(result){
+				let seq = $(this).parent().parent().prev().children().eq(2).val();
+				let board_Seq = $(this).parent().parent().prev().children().eq(3).val(); 
+				location.href="/delete.brd1_reply?seq="+seq+"&board_Seq="+board_Seq;	
+			}
+
 		})
 //			5. 댓글 좋아요 기능		
-		$(".good_path").on("click",function(){
-						
+		$(".good_path").on("click",function(){	
+			if(${nickname==null}){
+				alert("로그인 후 이용가능합니다.");
+				location.href="/index.jsp"
+			}
 			$.ajax({
 				url:"/good.brd1_reply",
 				data:{
@@ -313,7 +390,10 @@ div {
 //			6. 댓글 좋아요 취소 기능
 
 			$(".goodcencel_path").on("click",function(){
-						
+				if(${nickname==null}){
+					alert("로그인 후 이용가능합니다.");
+					location.href="/index.jsp"
+				}		
 			$.ajax({
 				url:"/cancelGood.brd1_reply",
 				data:{
